requirements.txt
langchain
langchain-community
langchain-chroma
langchain-groq
streamlit
pypdf
python-dotenv
faker
